pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.6",
    "django-environ>=0.11.2",
    "psycopg2-binary>=2.9.9",
    "gunicorn>=21.2.0",
    "django-rest>=0.8.7",
    "django-allauth[socialaccount]>=65.11.2",
    "djangorestframework-simplejwt>=5.5.1",
    "dj-rest-auth[with-social]>=7.0.1",
    "django-oauth-toolkit>=3.0.1",
    "django-cors-headers>=4.9.0",
    "django-extensions>=4.1",
]

# Dev-only tools (not installed in runtime images)
[dependency-groups]
dev = [
    "pre-commit",
]

# Tool configurations (picked up by hooks and local runs)
[tool.black]
line-length = 88
target-version = ["py313"]

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203","W503"]
exclude = [".venv","venv","build","dist","node_modules","migrations","__pycache__"]
